extracted settingsBar into its own component
